src/cloud_functions/function2/requirements.txt
google-cloud-bigquery==2.26.0
Flask==1.1.2
